+++ lib/Driver/ServerDriver.php
@@ -14,7 +14,7 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  *
- * @version    0.3.0
+ * @version    0.4.0
  * @copyright  2017-2020 Kristuff
  */
 
@@ -65,7 +65,7 @@ abstract class ServerDriver extends DatabaseDriver
      *
      * @access public
      * @param  string   $databaseName   The database name.
-     * @param  string   $owner          (optional) The database owner. This parameter is honored in pgsql only.
+     * @param  string   $owner          The database owner. This parameter is honored in pgsql only.
      *
      * @return bool     True if the database has been created, otherwise false.
      */