+++ lib/Exception/PatabaseException.php
@@ -14,7 +14,7 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  *
- * @version    0.3.0
+ * @version    0.4.0
  *
  * @copyright  2017-2020 Kristuff
 */
@@ -26,4 +26,6 @@ namespace Kristuff\Patabase\Exception;
  *
  * Base class for library custom Exceptions
  */
-class PatabaseException extends \Exception { }
+class PatabaseException extends \Exception 
+{
+}