+++ lib/Query/Insert.php
@@ -14,23 +14,18 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  *
- * @version    0.3.0
+ * @version    0.4.0
  * @copyright  2017-2020 Kristuff
  */
 
 namespace Kristuff\Patabase\Query;
-
-use Kristuff\Patabase;
-use Kristuff\Patabase\Query;
-use Kristuff\Patabase\Query\InsertBase;
-use Kristuff\Patabase\Database;
 
 /**
  * Class Insert
  *
  * Represents a [INSERT INTO] SQL query
  */
-class Insert extends InsertBase
+class Insert extends \Kristuff\Patabase\Query\InsertBase
 {
     /**
      * Build the SQL INSERT query