+++ lib/Query/InsertBase.php
@@ -14,7 +14,7 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  *
- * @version    0.3.0
+ * @version    0.4.0
  * @copyright  2017-2020 Kristuff
  */
 
@@ -102,7 +102,7 @@ abstract class InsertBase extends QueryBuilder
     public function bindValues()
     {
         foreach ($this->parameters as $key => $val) {
-            $arg = self::getArgName($key);
+            $arg = $this->getArgName($key);
             $this->pdoParameters[$arg] = $val;
         }
         parent::bindValues();