+++ lib/Query/Update.php
@@ -14,17 +14,15 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  *
-* @version    0.3.0
+* @version    0.4.0
  *
  * @copyright  2017-2020 Kristuff
  */
 
 namespace Kristuff\Patabase\Query;
 
-use Kristuff\Patabase;
 use Kristuff\Patabase\Query;
 use Kristuff\Patabase\Query\InsertBase;
-use Kristuff\Patabase\Database;
 
 /**
  * Class Update