+++ lib/Query/Where.php
@@ -14,16 +14,14 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  *
- * @version    0.3.0
+ * @version    0.4.0
  * @copyright  2017-2020 Kristuff
  */
 
 namespace Kristuff\Patabase\Query;
 
-use Kristuff\Patabase;
 use Kristuff\Patabase\Query\QueryBuilder;
 use Kristuff\Patabase\Query\QueryFilter;
-use Kristuff\Patabase\Driver\DatabaseDriver;
 
 /**
  * Class Where