+++ tests/pgsql/PgsqlInjectionTest.php
@@ -1,0 +1,37 @@
+<?php
+
+require_once __DIR__.'/../../vendor/autoload.php';
+require_once __DIR__.'/../base/DatabaseInjectionTest.php';
+
+use Kristuff\Patabase\Database;
+use Kristuff\Patabase\Server;
+
+class PgsqlInjectionTest extends DatabaseInjectionTest
+{
+    public static function setUpBeforeClass() : void
+    {   
+      
+        $settings = [
+            'driver'    => 'pgsql', 
+            'hostname'  => 'localhost', 
+            'username'  => 'postgres',
+            'password'  => ''
+        ];
+      
+        $srv = new Server($settings);
+        $srv->createDatabaseAndUser('patabaseTestInjection','tutu', 'pass');
+     
+        $dbsettings = [
+            'driver'    => 'pgsql', 
+            'hostname'  => 'localhost', 
+            'database'  => 'patabaseTestInjection',
+            'username'  => 'tutu',
+            'password'  => 'pass'
+        ];
+        self::$db = new Database($dbsettings);
+        self::createTables();
+  
+    }
+
+
+}